Refactor trigger removal to only revoke allowance if last trigger
Moved `CurrentStopLoss` from types to trigger-encoders. Introduced a conditional removal of the allowance. The modification checks if the current trigger is the last one in the account before revoking the allowance. This optimizes gas usage and improves performance when multiple triggers are set for the same account.
