+++ apps/summerfi-api/lib/setup-trigger-function/src/services/against-position-validators/aave-auto-sell-validator.ts
@@ -13,7 +13,7 @@ import {
 import { z } from 'zod'
 import { GetTriggersResponse } from '@summerfi/serverless-contracts/get-triggers-response'
 import { chainIdSchema, safeParseBigInt } from '@summerfi/serverless-shared'
-import { CurrentStopLoss } from '../types/current-stop-loss'
+import { CurrentStopLoss } from '../trigger-encoders'
 
 const paramsSchema = z.object({
   position: positionSchema,
