+++ apps/summerfi-api/lib/setup-trigger-function/src/services/get-aave-auto-sell-service-container.ts
@@ -94,6 +94,7 @@ export const getAaveAutoSellServiceContainer: (
         ? {
             triggerData: currentAutoSell.triggerData as `0x${string}`,
             id: safeParseBigInt(currentAutoSell.triggerId) ?? 0n,
+            triggersOnAccount: triggers.triggersCount,
           }
         : undefined
