+++ apps/summerfi-api/lib/setup-trigger-function/src/services/trigger-encoders/encode-aave-auto-buy.ts
@@ -59,7 +59,11 @@ export const encodeAaveAutoBuy = (
     ? encodeFunctionData({
         abi: automationBotAbi,
         functionName: 'removeTriggers',
-        args: [[currentTrigger.id], [currentTrigger.triggerData], true],
+        args: [
+          [currentTrigger.id],
+          [currentTrigger.triggerData],
+          currentTrigger.triggersOnAccount === 1, //remove allowance only if it's the last trigger
+        ],
       })
     : undefined
