+++ apps/summerfi-api/lib/setup-trigger-function/src/services/trigger-encoders/encode-aave-trailing-stop-loss.ts
@@ -73,7 +73,11 @@ export const encodeAaveTrailingStopLoss = (
     ? encodeFunctionData({
         abi: automationBotAbi,
         functionName: 'removeTriggers',
-        args: [[currentTrigger.id], [currentTrigger.triggerData], true],
+        args: [
+          [currentTrigger.id],
+          [currentTrigger.triggerData],
+          currentTrigger.triggersOnAccount === 1, // remove allowance only if it's the last trigger
+        ],
       })
     : undefined
