+++ apps/summerfi-api/lib/setup-trigger-function/src/services/trigger-encoders/encode-spark-auto-sell.ts
@@ -67,7 +67,11 @@ export const encodeSparkAutoSell = (
     ? encodeFunctionData({
         abi: automationBotAbi,
         functionName: 'removeTriggers',
-        args: [[currentTrigger.id], [currentTrigger.triggerData], true],
+        args: [
+          [currentTrigger.id],
+          [currentTrigger.triggerData],
+          currentTrigger.triggersOnAccount === 1, // remove allowance only if it's the last trigger
+        ],
       })
     : undefined
