+++ apps/summerfi-api/lib/setup-trigger-function/src/services/trigger-encoders/encode-spark-stop-loss.ts
@@ -69,7 +69,11 @@ export const encodeSparkStopLoss = (
     ? encodeFunctionData({
         abi: automationBotAbi,
         functionName: 'removeTriggers',
-        args: [[currentTrigger.id], [currentTrigger.triggerData], true],
+        args: [
+          [currentTrigger.id],
+          [currentTrigger.triggerData],
+          currentTrigger.triggersOnAccount === 1, //remove allowance only if it's the last trigger
+        ],
       })
     : undefined
