+++ apps/summerfi-api/lib/setup-trigger-function/src/services/trigger-encoders/encode-spark-trailing-stop-loss.ts
@@ -74,7 +74,11 @@ export const encodeSparkTrailingStopLoss = (
     ? encodeFunctionData({
         abi: automationBotAbi,
         functionName: 'removeTriggers',
-        args: [[currentTrigger.id], [currentTrigger.triggerData], true],
+        args: [
+          [currentTrigger.id],
+          [currentTrigger.triggerData],
+          currentTrigger.triggersOnAccount === 1, // remove allowance only if it's the last trigger
+        ],
       })
     : undefined
