+++ apps/summerfi-api/lib/setup-trigger-function/src/services/trigger-encoders/types.ts
@@ -1,3 +1,5 @@
+import { LTV, Price } from '~types'
+
 export type TriggerTransactions = {
   encodedTriggerData: `0x${string}`
   upsertTrigger: `0x${string}`
@@ -7,4 +9,10 @@ export type TriggerTransactions = {
 export type CurrentTriggerLike = {
   id: bigint
   triggerData: `0x${string}`
+  triggersOnAccount: number
+}
+
+export interface CurrentStopLoss extends CurrentTriggerLike {
+  executionPrice: Price
+  executionLTV: LTV
 }
