--- apps/summerfi-api/lib/setup-trigger-function/src/services/types/current-stop-loss.ts
@@ -1,8 +1,0 @@
-import { LTV, Price } from '~types'
-
-export interface CurrentStopLoss {
-  id: bigint
-  triggerData: `0x${string}`
-  executionPrice: Price
-  executionLTV: LTV
-}
